requirements.txt
altgraph
astroid==3.3.8
black==25.1.0
cfgv==3.4.0
click==8.1.8
dill==0.3.9
distlib==0.3.9
filelock==3.17.0
future
identify==2.6.6
isort==6.0.0
macholib
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pre_commit==4.1.0
pylint==3.3.4
PyYAML==6.0.2
six
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
virtualenv==20.29.1
requests
beautifulsoup4==4.12.2
sqlalchemy==2.0.25
pydantic==2.5.3
fastapi==0.110.0
uvicorn==0.29.0
pytest==8.0.0
httpx==0.26.0
psycopg2-binary==2.9.9  # PostgreSQL support (optional)
python-dotenv==1.0.1  # For managing environment variables
urllib3==1.26.16
